requirements-openstack.txt
amqp==5.1.0
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
autopage==0.5.0
bcrypt==3.2.0
BenchExec==3.11
billiard==3.6.4.0
cchardet==2.1.7
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
clade==3.4.11
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cliff==3.10.1
cmd2==2.4.0
cryptography==36.0.1
debtcollector==2.5.0
decorator==5.1.1
Django==4.0.3
django-appconf==1.0.5
django-celery-results==2.3.0
django-compressor==3.1
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
dogpile.cache==1.1.5
graphviz==0.19.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
Jinja2==3.0.3
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==3.2.0
keystoneauth1==4.5.0
kombu==5.2.4
MarkupSafe==2.1.0
msgpack==1.0.3
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==0.61.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.5.0
oslo.config==8.8.0
oslo.context==4.1.0
oslo.i18n==5.1.0
oslo.log==4.7.0
oslo.serialization==4.3.0
oslo.utils==4.12.2
packaging==21.3
paramiko==2.10.2
pbr==5.8.1
pika==1.2.0
pluggy==1.0.0
ply==3.11
prettytable==3.2.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
pyinotify==0.9.6
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.1
pytest==7.1.0
python-cinderclient==8.3.0
python-dateutil==2.8.2
python-glanceclient==3.6.0
python-keystoneclient==4.4.0
python-neutronclient==7.8.0
python-novaclient==17.7.0
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
rcssmin==1.1.0
requests==2.27.1
requestsexceptions==1.4.0
rfc3986==2.0.0
rjsmin==1.2.0
setuptools-scm==6.4.2
simplejson==3.17.6
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stevedore==3.5.0
text-unidecode==1.3
tomli==2.0.1
ujson==5.1.0
urllib3==1.26.8
vine==5.0.0
warlock==1.3.3
wcwidth==0.2.5
wrapt==1.14.0
